split facility module into submodules in facility folder. updated exports and imports appropriately
